Properly save time index in the module cache
Prior to this, the `timestamp` field in the module cache improperly used
the module's `created_at` field to indicate the current index in the full
Forge module list. This naively seems like the right choice, because we
don't actually care about releases, we only care about a module's
existence. Unfortunately, the Forge API doesn't let you sort by
`created_by` so we cannot identify the creation of new modules by that
field either. Instead, we need to just download all modules with new
releases since the last run and just deduplicate the list.