Speed up module/class metric
This reorders some of the logic to take some of it outside the loop. On
a small dataset, this wasn't a big deal, but on very large
infrastructures, this could chew through a lot of time on duplicate
operations. This was a speed increase of 37% already on my tiny test
environment, I expect that to be exponentially larger on huge
environments.