Load Ruby's CA certificates instead of using httpclient defaults
The `httpclient` gem uses a CA certificate bundle from 2015 (as of
version 2.8.3), see https://github.com/nahi/httpclient/issues/445. When
the Let's Encrypt root cert expired recently, this gem started failing
SSL handshakes, because the CA bundle it was using did not contain an
alternate trust path.

Calling `set_default_paths` on the client's SSL store causes it to load
the certs shipped with the Ruby runtime being used to execute dropsonde.
In most cases, these certs will be more up-to-date, and should allow the
connection to succeed. This is definitely true when running with
puppet-agent's Ruby.

With Windows system Ruby, this won't be sufficient. Recommended
workaround is to use puppet-agent's Ruby on Windows if possible.