(SERVER-3079) Allow overriding Puppet's base settings
When Dropsonde is run from within Puppet Server, Puppet calculates its
default settings values for the non-root 'puppet' user. These paths are
incorrect and need to be overridden. Because we need to initialize
Puppet's settings before parsing commandline args, this commit adds
environment variables that can be set to override the relevant values.